move/#34: 파일위치 정리
