+++ app/src/main/java/com/example/togedy_android/presentation/calendar/addPersonalSchedule/AddPersonalScheduleViewModel.kt
@@ -1,0 +1,11 @@
+package com.example.togedy_android.presentation.calendar.addPersonalSchedule
+
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class AddPersonalScheduleViewModel @Inject constructor(
+
+) {
+
+}
